Move old files and uis of samples to legacy dir
