+++ nicos_ess/loki/gui/legacy/ui_files/sampleconf.ui
