+++ nicos_ess/loki/gui/legacy/ui_files/sampleconf_one.ui
